Add a mark-needs-review command by introducing a concept of subcommands
The idea with the mark-needs-review command is that evaluate/reivew command should be updated to only reivew strings marked as "needs review" and change them to translated if the evaluation result is good
